add sqlitedict for data persistence
I looked at several options for adding data storarge, including: sqlite,
leveldb, and lmdb. I decided immediately that I wanted a schemeless
solution because while SQL is powerful, it is a lot more complicated to
set up and overkill for most projects.

lmdb was very fast and had automatic concurrency management. But the
database files are not automatically compacted and compacting them is
difficult.

leveldb is fast as well and had an easier Python API (for example,
iterating). And it has automatic compaction. It does not support
concurrency natively. Both leveldb and lmdb store the data in a
directory comprising multiple files.

sqlitedict uses sqlite with a simple key/value schema. The API provides
automatic serialization for most Python data types and simple assignment
to set objects. sqlitedict does not provide automatic vacuuming, but it
is extremely easy to add using sqlite's VACUUM command. sqlitedict
stores its data in a single sqlite database file, which is more
convenient than the leveldb approach. Additionally, sqlite has an
incredible history of stability.

While not as fast as leveldb, I felt sqlitedict offered the best
balance as a starting point for most projects. Of course, using sqlite
or leveldb or whatever is easy enough for projects that don't want
sqlitedict.